Initial implementation of the final step towards cross-version support. This current state only supports message-based data sharing - it specifically does not support dstore at this time as the relevant components need to be implemented.
Fix the resolve_peers and resolve_nodes APIs. This highlighted the need for gds/fetch API to handle info arrays, so update that framework API a bit. Reduce the number of params in the corresponding macro to make it easier to use.

Cleanup a bit so it passes tests

Fix the spawn test

Lots of cleanups - still not quite correct. Works okay on first pass thru simpclient, but get an error on second call to fence

Cleanup the scope argument for PMIx_Get, and process the info array for directives

Correctly add num local procs to the collectives tracker. Fix double-release on buffer. Update simpft test as notifications do not come back to the one generating the event.

Few more minor cleanups

Fix the atomics for embedded scenarios. Minor change to one test

Backport some fixes from OMPI

Final cleanups - operating okay in OMPI now

Sync with latest master

Signed-off-by: Ralph Castain <rhc@open-mpi.org>

Properly initialize the notification ring buffer in pmix client

Signed-off-by: Ralph Castain <rhc@open-mpi.org>

Fix a couple of naming conflicts when embedded

Signed-off-by: Ralph Castain <rhc@open-mpi.org>

Silence common variable and unused variable warnings

Signed-off-by: Ralph Castain <rhc@open-mpi.org>

Do not send event notifications if a process disconnects after calling finalize - this is normal termination and not an error

Signed-off-by: Ralph Castain <rhc@open-mpi.org>

Revise the GDS to include a new "cache_job_info" API that allows the server to put the job info into a fetchable place while waiting for procs from that nspace to connect. This allows test-resolve-peers to successfully complete.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>